more Firebase and KitchenNode; still need to do orderNumbers
